generative_retriever/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
farm-haystack[elasticsearch]==1.16.1
langchain==0.0.174
sentence-transformers==2.2.2
tiktoken==0.4.0
redis==4.5.5